Improved some unit tests
